Init Project with working prototype from backup
After having removed all, added a gitignore and then commit that gitignore, i am re-adding the 78 files related to the project that are NOT declared in the git ignore, not the 18thousand files that shouldn't be tracked. The libraries should be handled either by unity itself, or added as an external reference.